Lots of BnP code clean up.
Deleted the following files completely (dead or duplicate code):

contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/AvroStoreBuilderReducerPerBucket.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/HadoopStoreBuilderReducerPerBucket.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/IdentityJsonMapper.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/IdentityJsonReducer.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/JobState.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/VoldemortStoreBuilderMapper.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonConfigurable.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonDeserializerComparator.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonMapper.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonOutputCollector.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonReducer.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/serialization/JsonSequenceFileOutputFormat.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/utils/EmailMessage.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/utils/KeyValuePartitioner.java
contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/utils/MapperKeyValueWriter.java

Other changes:

- Changed some hard-coded strings to constants.
- Refactored common Mapper and Collector code via the new BuildAndPushMapper
  and AbstractCollectorWrapper classes.
- Refactored common 'reducer.per.bucket' code. HadoopStoreWriterPerBucket
  offers a superset of the HadoopStoreWriter's functionality, so it's not
  worth keeping the latter at all.
