+++ contrib/hadoop-store-builder/src/java/voldemort/store/readonly/disk/KeyValueWriter.java
@@ -22,7 +22,7 @@ import java.util.Iterator;
 import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.Reporter;
 
-// Interface used by reducers to layout the datqa on disk
+// Interface used by reducers to layout the data on disk
 public interface KeyValueWriter<K, V> {
 
     /*
