+++ contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/AbstractStoreBuilderConfigurable.java
@@ -25,16 +25,17 @@ import org.apache.hadoop.mapred.JobConf;
 import voldemort.VoldemortException;
 import voldemort.cluster.Cluster;
 import voldemort.store.StoreDefinition;
+import voldemort.store.readonly.ReadOnlyUtils;
+import voldemort.store.readonly.mr.azkaban.VoldemortBuildAndPushJob;
+import voldemort.utils.ByteUtils;
 import voldemort.xml.ClusterMapper;
 import voldemort.xml.StoreDefinitionsMapper;
 
 /**
- * A base class with basic configuration values shared by all the mapper,
- * reducer, and partitioner
- * 
- * 
+ * A base class with basic configuration values and utility functions
+ * shared by the mapper, reducer, and partitioner
  */
-public class AbstractStoreBuilderConfigurable {
+abstract public class AbstractStoreBuilderConfigurable {
 
     private int numChunks;
     private Cluster cluster;
@@ -49,12 +50,12 @@ public class AbstractStoreBuilderConfigurable {
             throw new IllegalStateException("Expected to find only a single store, but found multiple!");
         this.storeDef = storeDefs.get(0);
 
-        this.numChunks = conf.getInt("num.chunks", -1);
+        this.numChunks = conf.getInt(VoldemortBuildAndPushJob.NUM_CHUNKS, -1);
         if(this.numChunks < 1)
-            throw new VoldemortException("num.chunks not specified in the job conf.");
+            throw new VoldemortException(VoldemortBuildAndPushJob.NUM_CHUNKS + " not specified in the job conf.");
 
-        this.saveKeys = conf.getBoolean("save.keys", false);
-        this.reducerPerBucket = conf.getBoolean("reducer.per.bucket", false);
+        this.saveKeys = conf.getBoolean(VoldemortBuildAndPushJob.SAVE_KEYS, false);
+        this.reducerPerBucket = conf.getBoolean(VoldemortBuildAndPushJob.REDUCER_PER_BUCKET, false);
     }
 
     @SuppressWarnings("unused")
@@ -92,4 +93,36 @@ public class AbstractStoreBuilderConfigurable {
         return this.numChunks;
     }
 
+    public int getPartition(byte[] key,
+                            byte[] value,
+                            int numReduceTasks) {
+        try {
+            int partitionId = ByteUtils.readInt(value, ByteUtils.SIZE_OF_INT);
+            int chunkId = ReadOnlyUtils.chunk(key, getNumChunks());
+            if(getSaveKeys()) {
+                int replicaType = (int) ByteUtils.readBytes(value,
+                                                            2 * ByteUtils.SIZE_OF_INT,
+                                                            ByteUtils.SIZE_OF_BYTE);
+                if(getReducerPerBucket()) {
+                    return (partitionId * getStoreDef().getReplicationFactor() + replicaType)
+                            % numReduceTasks;
+                } else {
+                    return ((partitionId * getStoreDef().getReplicationFactor() * getNumChunks())
+                            + (replicaType * getNumChunks()) + chunkId)
+                            % numReduceTasks;
+                }
+            } else {
+                if(getReducerPerBucket()) {
+                    return partitionId % numReduceTasks;
+                } else {
+                    return (partitionId * getNumChunks() + chunkId) % numReduceTasks;
+                }
+            }
+        } catch (Exception e) {
+            throw new VoldemortException("Caught exception in getPartition()!" +
+                                         " key: " + ByteUtils.toHexString(key) +
+                                         ", value: " + ByteUtils.toHexString(value) +
+                                         ", numReduceTasks: " + numReduceTasks, e);
+        }
+    }
 }
