+++ contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/HadoopStoreBuilder.java
@@ -58,6 +58,7 @@ import voldemort.store.readonly.checksum.CheckSum;
 import voldemort.store.readonly.checksum.CheckSum.CheckSumType;
 import voldemort.store.readonly.checksum.CheckSumMetadata;
 import voldemort.store.readonly.disk.KeyValueWriter;
+import voldemort.store.readonly.mr.azkaban.VoldemortBuildAndPushJob;
 import voldemort.utils.ByteUtils;
 import voldemort.utils.Utils;
 import voldemort.xml.ClusterMapper;
@@ -171,18 +172,14 @@ public class HadoopStoreBuilder {
             conf.set("cluster.xml", new ClusterMapper().writeCluster(cluster));
             conf.set("stores.xml",
                      new StoreDefinitionsMapper().writeStoreList(Collections.singletonList(storeDef)));
-            conf.setBoolean("save.keys", saveKeys);
-            conf.setBoolean("reducer.per.bucket", reducerPerBucket);
+            conf.setBoolean(VoldemortBuildAndPushJob.SAVE_KEYS, saveKeys);
+            conf.setBoolean(VoldemortBuildAndPushJob.REDUCER_PER_BUCKET, reducerPerBucket);
             if(!isAvro) {
                 conf.setPartitionerClass(HadoopStoreBuilderPartitioner.class);
                 conf.setMapperClass(mapperClass);
                 conf.setMapOutputKeyClass(BytesWritable.class);
                 conf.setMapOutputValueClass(BytesWritable.class);
-                if(reducerPerBucket) {
-                    conf.setReducerClass(HadoopStoreBuilderReducerPerBucket.class);
-                } else {
-                    conf.setReducerClass(HadoopStoreBuilderReducer.class);
-                }
+                conf.setReducerClass(HadoopStoreBuilderReducer.class);
             }
             conf.setInputFormat(inputFormatClass);
             conf.setOutputFormat(SequenceFileOutputFormat.class);
@@ -248,7 +245,7 @@ public class HadoopStoreBuilder {
                     numReducers = cluster.getNumberOfPartitions() * numChunks;
                 }
             }
-            conf.setInt("num.chunks", numChunks);
+            conf.setInt(VoldemortBuildAndPushJob.NUM_CHUNKS, numChunks);
             conf.setNumReduceTasks(numReducers);
 
             if(isAvro) {
@@ -271,13 +268,7 @@ public class HadoopStoreBuilder {
                                                            Schema.create(Schema.Type.BYTES)));
 
                 AvroJob.setMapperClass(conf, mapperClass);
-
-                if(reducerPerBucket) {
-                    conf.setReducerClass(AvroStoreBuilderReducerPerBucket.class);
-                } else {
-                    conf.setReducerClass(AvroStoreBuilderReducer.class);
-                }
-
+                conf.setReducerClass(AvroStoreBuilderReducer.class);
             }
 
             logger.info("Number of chunks: " + numChunks + ", number of reducers: " + numReducers
