+++ contrib/hadoop-store-builder/src/java/voldemort/store/readonly/mr/azkaban/VoldemortStoreBuilderJob.java
@@ -30,15 +30,12 @@ import voldemort.store.StoreDefinition;
 import voldemort.store.readonly.checksum.CheckSum.CheckSumType;
 import voldemort.store.readonly.mr.AvroStoreBuilderMapper;
 import voldemort.store.readonly.mr.HadoopStoreBuilder;
-import voldemort.store.readonly.mr.VoldemortStoreBuilderMapper;
+import voldemort.store.readonly.mr.JsonStoreBuilderMapper;
 import voldemort.store.readonly.mr.serialization.JsonSequenceFileInputFormat;
 import voldemort.utils.Props;
 
 /**
  * Build a voldemort store from input data.
- * 
- * @author jkreps
- * 
  */
 public class VoldemortStoreBuilderJob extends AbstractHadoopJob {
 
@@ -191,7 +188,7 @@ public class VoldemortStoreBuilderJob extends AbstractHadoopJob {
     }
 
     public void run() throws Exception {
-        JobConf configuration = this.createJobConf(VoldemortStoreBuilderMapper.class);
+        JobConf configuration = this.createJobConf();
 
         Class mapperClass;
         Class<? extends InputFormat> inputFormatClass;
@@ -202,14 +199,12 @@ public class VoldemortStoreBuilderJob extends AbstractHadoopJob {
             configuration.set("avro.rec.schema", conf.getRecSchema());
             configuration.set("avro.key.schema", conf.getKeySchema());
             configuration.set("avro.val.schema", conf.getValSchema());
-
-            configuration.set("avro.key.field", conf.getKeyField());
-            configuration.set("avro.value.field", conf.getValueField());
-
+            configuration.set(VoldemortBuildAndPushJob.AVRO_KEY_FIELD, conf.getKeyField());
+            configuration.set(VoldemortBuildAndPushJob.AVRO_VALUE_FIELD, conf.getValueField());
             mapperClass = AvroStoreBuilderMapper.class;
             inputFormatClass = AvroInputFormat.class;
         } else {
-            mapperClass = VoldemortStoreBuilderMapper.class;
+            mapperClass = JsonStoreBuilderMapper.class;
             inputFormatClass = JsonSequenceFileInputFormat.class;
         }
